+++ magentoerpconnect/magento_model.py
@@ -382,8 +382,10 @@ class magento_store(orm.Model):
             'Create invoice on action',
             required=True,
             help="Should the invoice be created in Magento "
-                 "when it is validated or when it is paid in OpenERP?" 
-                  "This only takes effect if the SO-related Payment Method is not giving a Option for this by itself. (See Payment Methods)"),
+                 "when it is validated or when it is paid in OpenERP?\n"
+                  "This only takes effect if the sales order's related "
+                  "payment method is not giving an option for this by "
+                  "itself. (See Payment Methods)"),
     }
 
     _defaults = {
