+++ magentoerpconnect/payment_invoice.py
@@ -1,13 +1,17 @@
-from osv import osv, fields
+# -*- coding: utf-8 -*-
+from openerp.osv import orm, fields
 
-class payment_invoice(osv.Model):
-	_inherit = "payment.method"
-	
-	_columns = {
-		'create_invoice_on': fields.selection(
+
+class payment_invoice(orm.Model):
+    _inherit = "payment.method"
+
+    _columns = {
+        'create_invoice_on': fields.selection(
             [('open', 'Validate'),
              ('paid', 'Paid')],
             'Create invoice on action',
             help="Should the invoice be created in Magento "
-                 "when it is validated or when it is paid in OpenERP? If nothing is set, Create invoice on action falls back to Store Options."),
-		}
+                 "when it is validated or when it is paid in OpenERP?\n"
+                 "If nothing is set, the option falls back to the same option "
+                 "on the Magento store related to the sales order."),
+    }
