+++ magentoerpconnect/tests/__init__.py
@@ -23,7 +23,6 @@ import test_synchronization
 import test_address_book
 import test_export_invoice
 import test_import_product_image
-import test_export_invoice_jan
 fast_suite = [
 ]
